+++ src/main/java/org/betonquest/betonquest/api/bukkit/event/npc/NpcInteractEvent.java
@@ -4,7 +4,7 @@ import org.betonquest.betonquest.api.bukkit.event.ProfileEvent;
 import org.betonquest.betonquest.api.profile.OnlineProfile;
 import org.betonquest.betonquest.api.quest.npc.Npc;
 import org.betonquest.betonquest.id.NpcID;
-import org.betonquest.betonquest.objective.EntityInteractObjective.Interaction;
+import org.betonquest.betonquest.quest.objective.interact.Interaction;
 import org.bukkit.event.Cancellable;
 import org.bukkit.event.HandlerList;