+++ src/main/java/org/betonquest/betonquest/quest/objective/interact/Interaction.java
@@ -1,0 +1,19 @@
+package org.betonquest.betonquest.quest.objective.interact;
+
+/**
+ * The interaction type with an entity.
+ */
+public enum Interaction {
+    /**
+     * The player has to right-click the entity.
+     */
+    RIGHT,
+    /**
+     * The player has to left-click the entity.
+     */
+    LEFT,
+    /**
+     * The player can interact with the entity in any way.
+     */
+    ANY
+}