+++ src/test/java/org/betonquest/betonquest/quest/objective/interact/EntityInteractObjectiveTest.java
@@ -1,4 +1,4 @@
-package org.betonquest.betonquest.objective;
+package org.betonquest.betonquest.quest.objective.interact;
 
 import org.betonquest.betonquest.logger.util.BetonQuestLoggerService;
 import org.junit.jupiter.api.Test;